Completes benchmarking tests of 3D harmonic oscillators.  Currently settling on odeint openmp algebra + standard operations + openmp system() function, which seems to consistently be the fastest propagator.  I can't get the mkl operations to work with openmp algebra (though vector_algebra + mkl_operations + openmp system() function is only slightly slower).  For reference, propagating 1e6 3D harmonic oscillators through 1000 time steps (0:1:0.001) takes 14-15 seconds on my pitiful MacBook.

+++ molecool/src/debug/Profiler.h
@@ -285,7 +285,7 @@ namespace molecool {
 			auto highResStart = FloatingPointMicroseconds{ m_StartTimepoint.time_since_epoch() };
 			auto elapsedTime = std::chrono::time_point_cast<std::chrono::microseconds>(endTimepoint).time_since_epoch() - std::chrono::time_point_cast<std::chrono::microseconds>(m_StartTimepoint).time_since_epoch();
 
-			Profiler::get().gatherStats(m_Name, elapsedTime.count());
+			Profiler::get().gatherStats(m_Name, (double)elapsedTime.count());
 			if (MC_PROFILE_VERBOSE) {
 				Profiler::get().writeProfile({ m_Name, highResStart, elapsedTime, std::this_thread::get_id() });
 			}
